move tab component
