+++ src/components/WorkOrders/MobileWorkingWorkOrdersView/MobileWorkingWorkOrdersView.js
@@ -1,4 +1,4 @@
-import { useState, useEffect, useRef } from 'react'
+import { useState, useEffect, useRef, useRouter } from 'react'
 import { frontEndApiRequest } from '@/utils/frontEndApiClient/requests'
 import { beginningOfDay } from '@/utils/time'
 import { longMonthWeekday } from '@/utils/date'
@@ -8,15 +8,27 @@ import WarningInfoBox from '../../Template/WarningInfoBox'
 import Meta from '../../Meta'
 import { WorkOrder } from '../../../models/workOrder'
 import { MobileWorkingWorkOrderListItems } from './MobileWorkingWorkOrderListItems'
+import TabsVersionTwo from '../../TabsVersionTwo/Index'
 
 const SIXTY_SECONDS = 60 * 1000
 
 const MobileWorkingWorkOrdersView = ({ currentUser }) => {
+  const router = useRouter()
   const currentDate = beginningOfDay(new Date())
   const [visitedWorkOrders, setVisitedWorkOrders] = useState(null)
   const [sortedWorkOrders, setSortedWorkOrders] = useState(null)
 
   const [error, setError] = useState()
+
+  const titles = ['Current Work Orders', 'Past Work Orders']
+
+  const handleTabClick = (index) => {
+    index === 1 && router.push('/pastworkorders')
+  }
+
+  const ariaSelected = useMemo(() => {
+    return router.pathname === '/pastworkorders' ? 1 : 0
+  }, [router.pathname])
 
   const getOperativeWorkOrderView = async () => {
     setError(null)
@@ -93,6 +105,11 @@ const MobileWorkingWorkOrdersView = ({ currentUser }) => {
   return (
     <>
       <Meta title="Manage work orders" />
+      <TabsVersionTwo
+        titles={titles}
+        onTabChange={handleTabClick}
+        ariaSelected={ariaSelected}
+      />
       <div className="mobile-work-order-container">
         <h3 className="lbh-heading-h3">
           {longMonthWeekday(currentDate, { commaSeparated: false })}
